hotfix: migrations

+++ database/migrations/2017_03_02_035815_add_status_column_to_suburbs_table.php
@@ -13,9 +13,12 @@ class AddStatusColumnToSuburbsTable extends Migration
      */
     public function up()
     {
-        Schema::table('suburbs', function ($table) {
-            $table->integer('status')->default(true);
-        });
+        if (!Schema::hasColumn('suburbs', 'status')) {
+            Schema::table('suburbs', function ($table) {
+                $table->integer('status')->default(true);
+            });
+        }
+
     }
 
     /**
@@ -25,8 +28,11 @@ class AddStatusColumnToSuburbsTable extends Migration
      */
     public function down()
     {
-        Schema::table('suburbs', function ($table) {
-            $table->dropColumn('status');
-        });
+        if (Schema::hasColumn('suburbs', 'status')) {
+            Schema::table('suburbs', function ($table) {
+                $table->dropColumn('status');
+            });
+        }
+
     }
 }
